Workarounds to keep the site up
Dreamhosts have "upgraded" from debian to ubuntu. In the process
they appear to have either lost or relocated a bunch of core Perl
modules, some of which are used by the site.

More specifically DateTime and DBD::SQLite appear to be boned.

For the time being, the archive is down and no timestamps will be
displayed.